Move static methods out of utils into view model
